backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pytest>=8.2
pytest-django>=4.8
gunicorn==23.0.0
django-cors-headers==4.5.0
whitenoise==6.7.0
boto3>=1.26.0
python-dotenv>=1.0.0
